+++ tests/Unit/Drivers/GhasedakDriverTest.php
@@ -154,6 +154,21 @@ final class GhasedakDriverTest extends TestCase
         $this->callProtectedMethod($this->driver(), 'sendOtp', ['013', 'Otp message', '4567']);
     }
 
+    #[Test]
+    public function it_returns_credit_successfully(): void
+    {
+        Http::fake(['*' => Http::response(['Data' => ['Credit' => 1000]])]);
+
+        $credit = $this->driver()->credit();
+
+        $this->assertSame(1000, $credit);
+
+        Http::assertSent(fn (Request $request) => $request->url() === 'https://gateway.ghasedak.me/rest/api/v1/WebService/GetAccountInformation'
+                && $request->method() === 'GET'
+                && $request->hasHeader('ApiKey', 'sms_token')
+        );
+    }
+
     // -----------------
     // Helper Methods
     // -----------------