+++ CHANGELOG.md
@@ -11,9 +11,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - **Fara Payamak** provider.
 - **Ghasedak** provider.
-- **Limo SMS** provider.
 - **Behin Payam** provider.
 - **Asanak** provider.
+- `credit()` method in all providers to retrieve the current credit balance.
 
 ## [1.1.1] - 2025-08-12